requirements.txt
semantic-kernel>=1.28.1
mcp>=0.3.0
python-dotenv>=1.0.0
azure-identity>=1.14.0
requests>=2.31.0